Modificando ruta de interview manager
