+++ .env.example
@@ -1,0 +1,4 @@
+NEXT_PUBLIC_KC_ISSUER=keycloak_issuer_edit
+NEXT_PUBLIC_KC_CLIENT_ID=keycloack_client_edit
+NEXT_PUBLIC_KC_REDIRECT_URI=http://localhost:3000/callback
+NEXT_PUBLIC_KC_LOGOUT_REDIRECT_URI=http://localhost:3000
